fix/rename errors.ts to errorType.ts
